Feature - move correlation to logging
