+++ src/Arcus.WebApi.All/Arcus.WebApi.All.csproj
@@ -16,13 +16,11 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Arcus.WebApi.Correlation\Arcus.WebApi.Correlation.csproj" />
     <ProjectReference Include="..\Arcus.WebApi.Logging\Arcus.WebApi.Logging.csproj" />
     <ProjectReference Include="..\Arcus.WebApi.OpenApi.Extensions\Arcus.WebApi.OpenApi.Extensions.csproj" />
     <ProjectReference Include="..\Arcus.WebApi.Security.Authentication\Arcus.WebApi.Security.Authentication.csproj" />
     <ProjectReference Include="..\Arcus.WebApi.Security.Authorization\Arcus.WebApi.Security.Authorization.csproj" />
     <ProjectReference Include="..\Arcus.WebApi.Security\Arcus.WebApi.Security.csproj" />
-    <ProjectReference Include="..\Arcus.WebApi.Telemetry.Serilog\Arcus.WebApi.Telemetry.Serilog.csproj" />
   </ItemGroup>
 
 </Project>
