+++ src/Arcus.WebApi.Logging/Correlation/HttpCorrelationInfoAccessor.cs
@@ -3,7 +3,7 @@ using Arcus.Observability.Correlation;
 using GuardNet;
 using Microsoft.AspNetCore.Http;
 
-namespace Arcus.WebApi.Correlation
+namespace Arcus.WebApi.Logging.Correlation
 {
     /// <summary>
     /// Represents the correlation information on the current HTTP request, accessible throughout the application.
