+++ src/Arcus.WebApi.Logging/Correlation/IApplicationBuilderExtensions.cs
@@ -1,8 +1,7 @@
-﻿using System;
-using GuardNet;
+﻿using GuardNet;
 using Microsoft.AspNetCore.Builder;
 
-namespace Arcus.WebApi.Correlation 
+namespace Arcus.WebApi.Logging.Correlation 
 {
     /// <summary>
     /// Adds the <see cref="CorrelationMiddleware"/> to the application pipeline.
@@ -10,17 +9,6 @@ namespace Arcus.WebApi.Correlation
     // ReSharper disable once InconsistentNaming
     public static class IApplicationBuilderExtensions
     {
-        /// <summary>
-        /// Adds operation and transaction correlation to the application by using the <see cref="CorrelationMiddleware"/> in the request pipeline.
-        /// </summary>
-        [Obsolete("Correlation options is moved to 'Arcus.Observability.Correlation', use " + nameof(UseHttpCorrelation) + " instead")]
-        public static IApplicationBuilder UseCorrelation(this IApplicationBuilder applicationBuilder)
-        {
-            Guard.NotNull(applicationBuilder, nameof(applicationBuilder));
-
-            return UseHttpCorrelation(applicationBuilder);
-        }
-
         /// <summary>
         /// Adds operation and transaction correlation to the application by using the <see cref="CorrelationMiddleware"/> in the request pipeline.
         /// </summary>
