+++ src/Arcus.WebApi.sln
@@ -17,10 +17,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.OpenApi.Extens
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Security.Authentication", "Arcus.WebApi.Security.Authentication\Arcus.WebApi.Security.Authentication.csproj", "{88EADCAB-29BF-42F1-8748-646676898903}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Telemetry.Serilog", "Arcus.WebApi.Telemetry.Serilog\Arcus.WebApi.Telemetry.Serilog.csproj", "{C06FBEE1-A5D1-4D57-8090-781B26C519DB}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Correlation", "Arcus.WebApi.Correlation\Arcus.WebApi.Correlation.csproj", "{455086AD-C6AD-42FB-AA83-B294F64D9F26}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Tests.Integration", "Arcus.WebApi.Tests.Integration\Arcus.WebApi.Tests.Integration.csproj", "{257F0347-4F29-4386-B2A2-DEBF947C6E41}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Security.Authorization", "Arcus.WebApi.Security.Authorization\Arcus.WebApi.Security.Authorization.csproj", "{B3932833-DD99-4732-BCCB-3F8E53E8DA62}"
@@ -55,14 +51,6 @@ Global
 		{88EADCAB-29BF-42F1-8748-646676898903}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{88EADCAB-29BF-42F1-8748-646676898903}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{88EADCAB-29BF-42F1-8748-646676898903}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C06FBEE1-A5D1-4D57-8090-781B26C519DB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C06FBEE1-A5D1-4D57-8090-781B26C519DB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C06FBEE1-A5D1-4D57-8090-781B26C519DB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C06FBEE1-A5D1-4D57-8090-781B26C519DB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{455086AD-C6AD-42FB-AA83-B294F64D9F26}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{455086AD-C6AD-42FB-AA83-B294F64D9F26}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{455086AD-C6AD-42FB-AA83-B294F64D9F26}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{455086AD-C6AD-42FB-AA83-B294F64D9F26}.Release|Any CPU.Build.0 = Release|Any CPU
 		{257F0347-4F29-4386-B2A2-DEBF947C6E41}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{257F0347-4F29-4386-B2A2-DEBF947C6E41}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{257F0347-4F29-4386-B2A2-DEBF947C6E41}.Release|Any CPU.ActiveCfg = Release|Any CPU
